requirements.txt
# ─── Core & Web ──────────────────────────────────────────
flask>=3.0.0
twilio>=9.0.0
requests>=2.32.0
python-dotenv>=1.0.0
streamlit>=1.42.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20

# ─── AI & LLMs ──────────────────────────────────────────
openai>=1.61.0
google-generativeai>=0.8.0
langchain>=0.3.17
langchain-community>=0.3.16
langchain-openai>=0.3.4
langchain-chroma>=0.2.2

# ─── Vector Database & Data ─────────────────────────────
chromadb>=0.6.3
pypdf>=5.0.0
docx2txt>=0.8
tiktoken>=0.8.0
pydantic>=2.10.0

# ─── Google Services (Existing Bot) ─────────────────────
google-auth>=2.37.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.159.0
